Integrated FD and a console application. Make a simple view. Added interface for placer.

+++ Buddy.All/Buddy.Common/Vertex.cs
@@ -2,8 +2,8 @@
 {
     public struct Vertex
     {
-        public uint Id { get; set; }
+        public int Id { get; set; }
 
-        public double Radius { get; set; }
+        public float Radius { get; set; }
     }
 }
